Mehfils completely moved to their own views, urls, forms and template folder. (minor changes in abuse.py)
