+++ links/abuse.py
@@ -12,7 +12,7 @@ CORE_ENGLISH_ABUSE_WORDS = ['fuck','fck','fvck','fuk','fuq','phuck','phuk','phuc
 'f_ck','f.ck','porki-stan','porki.stan','porki_stan','fack','feck','porkies','p0rkies','pork!es','p0rk!es','s3x','s3ks','s-3-x','s_3_x','s.3.x','s.3_x',\
 's.3-x','s-3.x','s_3.x','s-3_x','sxsy','pron','romantic','rommantic','naughty','bobs','sixy','sx','x.xx','leon','whatapp','whatsapp','romentic','lesb',\
 'whatsap','sax','s@x','f.u.k','vegina','se_x','leso','vigina','pen@is','ediut','kiss','whtsap','b0o0b','b.o.o.b','sixe','shemaale','cex','iesbo','fu-ck',\
-'watsap']#'anus',butt','kok','anal','ana1','hot','fanny','qhus'
+'watsap','sperm']#'anus',butt','kok','anal','ana1','hot','fanny','qhus'
 
 
 CORE_URDU_ABUSE_WORDS = ['aunty','besharm','baysharm','phudd','phudi','phodi','lun','choot','gashti','gandu','luun','gshti','randi','rndi','hijra','phatti','nanga','nangi','phddi',\
@@ -30,10 +30,10 @@ CORE_URDU_ABUSE_WORDS = ['aunty','besharm','baysharm','phudd','phudi','phodi','l
 'k.a.n.j.a.r','f.u.d.d.i','f.u.d.i','p.o.d.i','p.u.d.d.a','p.u.d.d.d.a','k.h.u.s.r.a','yavai','y.a.v.a.i','y.u.v.a.i','l.u.l','c.h.o.d','s.u.a.r','p.d.d.a',\
 'k.h.a.s.s.i','k.h.a.s.i','hijry','hijre','h!jry','h!jre','phudy','phody','khassy','yavay','r@ndi','naangi','behn','bahn','bhen','-ki-','.ki.','_ki_','-ki_','_ki-',\
 '.ki_','_ki.','.ki-','-ki.','-ke-','.ke.','_ke_','-ke_','_ke-','.ke_','_ke.','.ke-','-ke.','gasshti','chuut','kanjjri','ghashti','ranndy','behen','behan',\
-'bahan','phuda','ph0di','hawas','havas','bgarat','lorro','bhan','bhund','phoudi','ph0da','g@ndo','puddii','puudii','chout','ganndi','ch0ut','hrami','kuuttii','l00ra','gaaaando',\
+'bahan','phuda','ph0d','hawas','havas','bgarat','lorro','bhan','bhund','g@ndo','phoudi','puddii','puudii','chout','ganndi','ch0ut','hrami','kuuttii','l00ra','gaaaando',\
 'marwani','tharki','th@rki','bgairat','raandi','gannd','knjr','rnnd','gndo','bhn','gnd','pyar','garum','chudakar','thko','phude','fhudiyo','chudwa','phudda',\
 'ghushti','jismani','chot','xhodo','thoko','ch0t','lu.n','hiijra','chadwa','pyase','l0n','ghand','gndo','dewani','dewana','uff','uuf','phuud','chdwa','ghsti',\
-'iun','knnjr','gnnd']#salle,lanti,lant,soor,
+'iun','knnjr','gnnd','gasht','gsht','barwy','l_u_n','ch0s']#salle,lanti,lant,soor,ph0di,ph0de
 
 
 CORE_POLITICALLY_SENSITIVE_WORDS = ['isis','hitler','hit1er','h!tler','h!t1er','h1t!er','hytler','hyt1er','hitla','hit1a','h!t1a','h!t!er','1s1s',\
@@ -45,18 +45,18 @@ MISC_POLITICALLY_SENSITIVE_WORDS = ['india','pakistan','america','amreeka','amri
 'mus1im','musl!m','mus1!m','yahudi','yahud!','mhb','m_h_b','m-h-b','m.h.b','m-h_b','m_h-b','m.h_b','m_h.b','m-h.b','m.h-b','mh.b','m.h@b']
 
 CORE_RELIGIOUSLY_SENSITIVE_WORDS = ['allah','bhagwan','kafir','qafir','kafr','qafr','kfir','qfir','kaphir','qaphir','kaf1r','kaf!r','a11ah','bhagvan','kafur',\
-'bhagwaan','bagwan','bagwaan','bagvaan','jesus','ahmedi','ahmadi','haram','haraam','haaraam','haraaam','ka-fir','ka-f!r','ka_f!r','ka-ph!r','ka-ph1r',\
+'bhagwaan','bagwan','bagwaan','bagvaan','jesus','ahmedi','ahmadi','haraam','haaraam','haraaam','ka-fir','ka-f!r','ka_f!r','ka-ph!r','ka-ph1r',\
 'qa-fir','qa-f!r','qa-f1r','athiest','atheist','atheism','athiesm','sipah','taliban','ta1iban','talibaan','tal!baan','tal!ban','talyban','talybaan','god',\
 'taaliban','taa1iban','allaah','blasphemy','kufr','qufr','sippa','s!ppa','s!pah','fatwa','fatva','ftwa','ftva','bhagvaan','k.a.f.i.r','q.a.f.i.r','k.a.f.r',\
 'k.f.i.r','q.a.f.r','q.f.i.r','k.a.p.h.i.r','q.a.p.h.i.r','k.a.f.!.r','q.a.f.!.r','k.a.f.u.r','kafer','qafer','kaffer','qaffer','kaffur','qaffur',\
-'h.a.r.a.m','h.a.r.a.a.m']
+'h.a.r.a.m','h.a.r.a.a.m']#haram
 
 MISC_SENSITIVE_WORDS = ['fahashi','jinsi','garam','kartoot','qartoot','kart00t','qart00t','kartut','qartut','karto0t','kart0ot','qart0ot','qarto0t',\
 'krtut','qrtut','kertut','kert00t','kertoot','qertoot','qert00t','idiot','idi0t','id!ot','id!0t','1diot','damn','kat!l','scandal','scandle','scendel',\
 'skendel','skendl','scandl','scendl','sharmnaak','sharamnaak','sharamnaaq','sharmnaaq','shrmnaak','shrmnaaq','jinsy','j!nsy','j!nsi','porn','p0rn',\
 'sharamnak','p-o-r-n','p_o_r_n','p.o.r.n','p-o_r-n','p_o-r_n','p-o_r_n','p_o_r-n']
 
-BANNED_FORGOTTEN_NICKS = ['fuck','sex','phudd','phudi','phodi','lun','choot','pussy','boob','xx','breast','penis','hitler','gashti','gandu','luun',\
+BANNED_FORGOTTEN_NICKS = ['fuck','sex','phudd','phudi','phodi','lun','choot','pussy','boob','xx','breast','penis','hitler','gasht','gandu','luun',\
 'mhb','allah','bhagwan','allaah','quran','jesus','Damadam-Feedback','Damadam-Admin']
 
 BANNED_NICKS = CORE_ENGLISH_ABUSE_WORDS+CORE_URDU_ABUSE_WORDS+CORE_POLITICALLY_SENSITIVE_WORDS+MISC_POLITICALLY_SENSITIVE_WORDS+\
