+++ unconnectedreddit/urls_mehfil.py
@@ -3,8 +3,10 @@ from django.conf.urls import patterns, url
 from django.contrib import admin
 from links.mehfil_views import GroupHelpView, ReinviteView, OpenGroupHelpView, ClosedGroupHelpView, MehfilView, AppointCaptainView, OwnerGroupOnlineKonView,\
 GroupOnlineKonView, DirectMessageCreateView, ClosedGroupCreateView, OpenGroupCreateView, InviteUsersToPrivateGroupView, DirectMessageView, ReinvitePrivateView,\
-InviteUsersToGroupView, GroupRankingView, appoint_pk, direct_message, leave_public_group, leave_private_group, left_private_group, mehfil_help, \
-process_public_group_invite, process_private_group_invite,invite_private
+InviteUsersToGroupView, GroupRankingView, GroupPageView, GroupTypeView, ChangeGroupRulesView, ChangePrivateGroupTopicView, ChangeGroupTopicView, \
+PublicGroupView, PrivateGroupView, appoint_pk, direct_message, leave_public_group, leave_private_group, left_private_group, mehfil_help, \
+process_public_group_invite,process_private_group_invite, invite_private, public_group_request_denied, public_group, left_public_group, \
+first_time_public_refresh, first_time_refresh, priv_group, del_public_group, kick_pk, groupreport_pk
 
 admin.autodiscover()
 
@@ -21,6 +23,7 @@ urlpatterns = patterns('',
 	url(r'^privatemehfil/invite/$', auth(InviteUsersToPrivateGroupView.as_view()), name='invite_private_group'),
 	url(r'^group/invite/$', auth(InviteUsersToGroupView.as_view()), name='invite'),
 	######################################### Mehfil creation #####################################
+	url(r'^group_type/$', auth(GroupTypeView.as_view()), name='group_type'),
 	url(r'^open_group/help/$', auth(OpenGroupHelpView.as_view()), name='open_group_help'),
 	url(r'^closed_mehfil/help/$', auth(ClosedGroupHelpView.as_view()), name='closed_group_help'),
 	url(r'^mehfil/help/$', auth(MehfilView.as_view()), name='mehfil_help'),
@@ -34,6 +37,8 @@ urlpatterns = patterns('',
 	url(r'^lvpgr/$', auth(leave_public_group), name='leave_public_group'),
 	url(r'^lvpg/$', auth(leave_private_group), name='leave_private_group'),
 	url(r'^leftgroup/$', auth(left_private_group), name='left_private_group'),
+	url(r'^leftpgroup/$', auth(left_public_group), name='left_public_group'),	
+	url(r'^dlpgr/(?P<pk>\d+)/(?P<unique>[\w.@+-]+)/(?P<private>\d+)/$', auth(del_public_group), name='del_public_group'),	
 	################################# Mehfil administration #######################################
 	url(r'^appoint/(?P<pk>\d+)/(?P<app>\d+)/$', auth(appoint_pk), name='appoint_pk'),
 	url(r'^appoint/$', auth(AppointCaptainView.as_view()), name='appoint'),
@@ -41,5 +46,26 @@ urlpatterns = patterns('',
 	url(r'^group/online_kon/$', auth(GroupOnlineKonView.as_view()), name='group_online_kon'),
 	########################## Popular mehfil list #########################
 	url(r'^popular_mehfil/$', auth(GroupRankingView.as_view()), name='group_ranking'),
-	
+	#################### Rendering list of all mehfils #####################
+	url(r'^group/$', auth(GroupPageView.as_view()), name='group_page'),
+	############################## Changing public and private mehfil topic ##############################
+	url(r'^group/change_rules/$', auth(ChangeGroupRulesView.as_view()), name='change_rules'),
+	url(r'^group/change_private_topic/$', auth(ChangePrivateGroupTopicView.as_view()), name='change_private_topic'),
+	url(r'^group/change_topic/$', auth(ChangeGroupTopicView.as_view()), name='change_topic'),
+	############################## Rendering and posting to private mehfil ##############################
+	url(r'^pmf/$', auth(priv_group), name='priv_group'),
+	url(r'^mehfil/private/$', auth(PrivateGroupView.as_view()), name='private_group_reply'),	
+	############################## Rendering and posting to public mehfil ##############################
+	url(r'^awami_mehfil_request_denied/$', auth(public_group_request_denied), name='public_group_request_denied'),
+	url(r'^mehfil/awami/$', auth(PublicGroupView.as_view()), name='public_group_reply'),
+	url(r'^mehfilawami/$', auth(public_group), name='public_group'),
+	url(r'^mehfilawami/$', auth(public_group), name='public_group'),
+	url(r'^mehfilawami/(?P<slug>[\w.@+-]+)/$', auth(public_group), name='public_group'),	
+
+	#################### Refreshing mehfils #####################
+	url(r'^ftpr/$', auth(first_time_public_refresh), name='first_time_public_refresh'),	
+	url(r'^ftr/(?P<unique>[\w.@+-]+)/$', auth(first_time_refresh), name='first_time_refresh'),	
+	############################## Mehfil punishments ##############################
+	url(r'^kick/$', auth(kick_pk), name='kick_pk'),	
+	url(r'^groupreport/$', auth(groupreport_pk), name="group_report_pk"),
 )
